+++ ChangeLog.md
@@ -1,4 +1,8 @@
-## 2018-01-14
+## 2018-01-14 (1/2)
+* Show CodePipeline stages with color indicating the current stage. Requires 
+Radiator Exposer update also.
+
+## 2018-01-14 (1/2)
 * Show commit author and message if known in pipeline status. Requires 
 Radiator Exposer update also.